fix: Updated samplers to properly assign sampling decisions based on appropriate context
