+++ lib/samplers/adaptive-sampler.js
@@ -108,19 +108,25 @@ class AdaptiveSampler extends Sampler {
   applySamplingDecision({ transaction, tracestate, partialType }) {
     if (!transaction) return
     transaction.partialType = partialType
-    if (tracestate) {
+    if (tracestate?.intrinsics) {
       // Explicitly set sampled and priority from tracestate intrinsics if available
-      transaction.sampled = tracestate?.intrinsics ? tracestate.isSampled : null
-      transaction.priority = tracestate?.intrinsics ? tracestate.priority : null
+      transaction.sampled = tracestate.isSampled
+      let initPriority = tracestate.priority
+      // If tracestate has intrinsics but lacks a priority we must generate a random priority
+      // if sampled, we also have to increment accordingly based on full/partial traces
+      if (!initPriority && tracestate.isSampled) {
+        initPriority = Sampler.generatePriority()
+        initPriority = Sampler.incrementPriority(initPriority, partialType)
+      }
+
+      transaction.priority = initPriority
       return
     }
 
     // If a tracestate is not defined, then do our normal priority calculation.
-
     const initPriority = Sampler.generatePriority()
     transaction.sampled = this.shouldSample(initPriority)
-    // only add 1 to priority when doing a full trace
-    transaction.priority = transaction.sampled && !partialType ? Sampler.incrementPriority(initPriority, 1) : initPriority
+    transaction.priority = transaction.sampled ? Sampler.incrementPriority(initPriority, partialType) : initPriority
   }
 
   /**
