+++ lib/samplers/ratio-based-sampler.js
@@ -26,8 +26,7 @@ class TraceIdRatioBasedSampler extends Sampler {
     transaction.partialType = partialType
     const initPriority = Sampler.generatePriority()
     transaction.sampled = this.shouldSample(transaction.traceId)
-    // only add 1 to priority when doing a full trace
-    transaction.priority = transaction.sampled && !partialType ? Sampler.incrementPriority(initPriority, 1) : initPriority
+    transaction.priority = transaction.sampled ? Sampler.incrementPriority(initPriority, partialType) : initPriority
   }
 
   /**
