+++ lib/samplers/sampler.js
@@ -45,11 +45,13 @@ class Sampler {
 
   /**
    * Used to increment priority and truncate to 6 decimal places.
+   * Full traces are incremented by 2 and partial traces are incremented by 1
    * @param {number} priority the current priority
-   * @param {number} increment the amount to increment by
+   * @param {string|null} partialType if a partial trace
    * @returns {number} the incremented priority
    */
-  static incrementPriority(priority, increment) {
+  static incrementPriority(priority, partialType) {
+    const increment = partialType ? 1 : 2
     const newPriority = priority + increment
     return ((newPriority * 1e6) | 0) / 1e6
   }
