+++ test/lib/cross_agent_tests/distributed_tracing/README.md
@@ -14,12 +14,19 @@ the agent under test. Here's what the various fields in each test case mean:
 | `account_id` | The account id the agent would receive on connect. |
 | `web_transaction` | Whether the transaction that's tested is a web transaction or not. |
 | `raises_exception` | Whether to simulate an exception happening within the transaction or not, resulting in a transaction error event. |
-| `root` | The sampler to use for transactions at the root of a trace. |
-| `remote_parent_sampled` | The sampler to use for transactions with a remote parent. |
-| `remote_parent_not_sampled` | The sampler to use for transactions with a remote parent that is not sampled. |
+| `distributed_tracing_enabled` | If `false`, then distributed tracing is disabled. If `true` or absent, then distributed tracing is enabled (default behavior). |
+| `full_granularity_enabled` | If `false`, then full granularity tracing is disabled. If `true` or absent, then full granularity is enabled (default behavior). |
+| `root` | The full granularity sampler to use for transactions at the root of a trace. |
+| `remote_parent_sampled` | The full granularity sampler to use for transactions with a remote parent that was sampled. |
+| `remote_parent_not_sampled` | The full granularity sampler to use for transactions with a remote parent that is not sampled. |
+| `force_adaptive_sampled` | The sampling decision to force on a transaction whenever the adaptive sampler is used. This applies to all adaptive samplers used in the test, whether they are the global sampler or an individual sampler instance. |
+| `full_granularity_ratio` | The ratio to use for all of the full granularity trace ID ratio samplers defined in the test. For testing purposes we are not defining different ratios for each trace ID ratio sampler instance. If that is necessary, we will need a different way to configure the ratios. |
+| `partial_granularity_enabled` | If `true`, then partial granularity is enabled. If `false` or absent, then partial granularity is disabled (default behavior). |
+| `partial_granularity_root` | The partial granularity sampler to use for root transactions. |
+| `partial_granularity_remote_parent_sampled` | The partial granularity sampler to use for transactions with a remote parent that was sampled. |
+| `partial_granularity_remote_parent_not_sampled` | The partial granularity sampler to use for transaction with a remote parent that was not sampled. |
+| `partial_granularity_ratio` | The partial granularity ratio to use for all the partial granularity ratio samplers defined in the test. As with `full_granularity_ratio` we're limiting these tests to have one ratio configured for all partial granularity samplers.| 
 | `expected_priority_between` | The inclusive range of the expected priority value on the generated transaction event. |
-| `force_adaptive_sampled` | The sampling decision to force on a transaction whenever the adaptive sampler is used. |
-| `ratio` | The ratio to use for all of the trace ID ratio samplers defined in the test. For testing purposes we are not defining different ratios for each trace ID ratio sampler instance. If that is necessary, we will need a different way to configure the ratios. |
 | `transport_type` | The transport type for the inbound request. |
 | `inbound_headers` | The headers you should mock coming into the agent. |
 | `outbound_payloads` | The exact/expected/unexpected values for outbound `w3c` headers. |
